+++ R/tune_nested.R
@@ -22,11 +22,7 @@
 #' @template param_store_benchmark_result
 #' @template param_store_models
 #' @template param_check_values
-#' @template param_allow_hotstart
-#' @template param_keep_hotstart_stack
-#' @template param_evaluate_default
 #' @template param_callbacks
-#' @template param_method
 #'
 #' @export
 #' @examples
@@ -45,12 +41,23 @@
 #'
 #' # Unbiased performance of the final model trained on the full data set
 #' rr$aggregate()
-tune_nested = function(tuner, task, learner, inner_resampling, outer_resampling, measure = NULL, term_evals = NULL, term_time = NULL, terminator = NULL, search_space = NULL, store_tuning_instance = TRUE, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, allow_hotstart = FALSE, keep_hotstart_stack = FALSE, evaluate_default = FALSE, callbacks = list(), method) {
-  if (!missing(method)) {
-    message("The `method` argument is deprecated and will be removed in a future release. Please use `tuner` instead.")
-    tuner = method
-  }
-
+tune_nested = function(
+  tuner,
+  task,
+  learner,
+  inner_resampling,
+  outer_resampling,
+  measure = NULL,
+  term_evals = NULL,
+  term_time = NULL,
+  terminator = NULL,
+  search_space = NULL,
+  store_tuning_instance = TRUE,
+  store_benchmark_result = TRUE,
+  store_models = FALSE,
+  check_values = FALSE,
+  callbacks = NULL
+  ) {
   assert_task(task)
   assert_resampling(inner_resampling)
   assert_resampling(outer_resampling)
@@ -68,9 +75,6 @@ tune_nested = function(tuner, task, learner, inner_resampling, outer_resampling,
     store_benchmark_result = store_benchmark_result,
     store_models = store_models,
     check_values = check_values,
-    allow_hotstart = allow_hotstart,
-    keep_hotstart_stack = keep_hotstart_stack,
-    evaluate_default = evaluate_default,
     callbacks = callbacks)
 
   resample(task, at, outer_resampling, store_models = TRUE)