+++ man-roxygen/param_host.R
@@ -1,0 +1,2 @@
+#' @param host (`character(1)`)\cr
+#' Worker is started on a local or remote host.