+++ man-roxygen/param_hotstart_threshold.R
@@ -1,0 +1,3 @@
+#' @param hotstart_threshold (`numeric(1)`)\cr
+#' Threshold for storing learners in the stack.
+#' If the value of the hotstart parameter is below this threshold, the learner is not added to the stack.