+++ man-roxygen/param_inst_async.R
@@ -1,0 +1,2 @@
+#' @param inst ([TuningInstanceAsyncSingleCrit] | [TuningInstanceAsyncMultiCrit] )\cr
+#'  The tuning instance.