+++ man-roxygen/param_packages.R
@@ -1,0 +1,3 @@
+#' @param packages (`character()`)\cr
+#'   Set of required packages.
+#'   Note that these packages will be loaded via [requireNamespace()], and are not attached.