+++ man-roxygen/param_start_workers.R
@@ -1,0 +1,2 @@
+#' @param start_workers (`logical(1)`)\cr
+#' If `TRUE`, the workers are started immediately with `future` on all available cores.