+++ tests/testthat/helper.R
@@ -4,6 +4,7 @@ library(mlr3misc)
 library(mlr3pipelines)
 library(paradox)
 library(R6)
+library(rush)
 
 lapply(list.files(system.file("testthat", package = "mlr3"), pattern = "^helper.*\\.[rR]", full.names = TRUE), source)
 lapply(list.files(system.file("testthat", package = "mlr3tuning"), pattern = "^helper.*\\.[rR]", full.names = TRUE), source)