+++ tests/testthat/teardown.R
@@ -1,3 +1,4 @@
 options(old_opts)
 lg_mlr3$set_threshold(old_threshold_mlr3)
 lg_bbotk$set_threshold(old_threshold_bbotk)
+lg_rush$set_threshold(old_threshold_rush)